Rearrange the files properly in the plugins directory and define the python scripts from the code in Calibration/HcalIsolatedTrackReco
