+++ Calibration/HcalIsolatedTrackReco/plugins/IsolatedEcalPixelTrackCandidateProducer.cc
@@ -22,16 +22,54 @@
 #include <memory>
 
 // user include files
-
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "DataFormats/Common/interface/Ref.h"
 #include "DataFormats/DetId/interface/DetId.h"
 #include "DataFormats/Math/interface/deltaR.h"
+#include "DataFormats/HcalIsolatedTrack/interface/IsolatedPixelTrackCandidate.h"
+#include "DataFormats/EcalRecHit/interface/EcalRecHitCollections.h"
+#include "DataFormats/HLTReco/interface/TriggerFilterObjectWithRefs.h"
 
-#include "Calibration/HcalIsolatedTrackReco/interface/IsolatedEcalPixelTrackCandidateProducer.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+
+#include "Geometry/CaloGeometry/interface/CaloGeometry.h"
+#include "Geometry/Records/interface/CaloGeometryRecord.h"
 
 //#define EDM_ML_DEBUG
+
+//
+// class decleration
+//
+
+class IsolatedEcalPixelTrackCandidateProducer : public edm::global::EDProducer<> {
+public:
+  explicit IsolatedEcalPixelTrackCandidateProducer(const edm::ParameterSet&);
+  ~IsolatedEcalPixelTrackCandidateProducer() override;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
+
+  const edm::EDGetTokenT<EcalRecHitCollection> tok_ee;
+  const edm::EDGetTokenT<EcalRecHitCollection> tok_eb;
+  const edm::EDGetTokenT<trigger::TriggerFilterObjectWithRefs> tok_trigcand;
+  const edm::ESGetToken<CaloGeometry, CaloGeometryRecord> tok_geom_;
+  const double coneSizeEta0_;
+  const double coneSizeEta1_;
+  const double hitCountEthrEB_;
+  const double hitEthrEB_;
+  const double fachitCountEE_;
+  const double hitEthrEE0_;
+  const double hitEthrEE1_;
+  const double hitEthrEE2_;
+  const double hitEthrEE3_;
+};
 
 IsolatedEcalPixelTrackCandidateProducer::IsolatedEcalPixelTrackCandidateProducer(const edm::ParameterSet& conf)
     : tok_ee(consumes<EcalRecHitCollection>(conf.getParameter<edm::InputTag>("EERecHitSource"))),
@@ -180,3 +218,8 @@ void IsolatedEcalPixelTrackCandidateProducer::produce(edm::StreamID,
 #endif
   iEvent.put(std::move(iptcCollection));
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+DEFINE_FWK_MODULE(IsolatedEcalPixelTrackCandidateProducer);
