--- Calibration/HcalIsolatedTrackReco/src/SealModule.cc
@@ -1,24 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/IsolatedPixelTrackCandidateProducer.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/IsolatedPixelTrackCandidateL1TProducer.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/IsolatedEcalPixelTrackCandidateProducer.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/EcalIsolatedParticleCandidateProducer.h"
-#include "RecoTracker/TkTrackingRegions/interface/TrackingRegionProducerFactory.h"
-#include "RecoTracker/TkTrackingRegions/interface/TrackingRegionProducer.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/SiStripRegFEDSelector.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/ECALRegFEDSelector.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/SubdetFEDSelector.h"
-#include "HITRegionalPixelSeedGenerator.h"
-#include "Calibration/HcalIsolatedTrackReco/interface/IPTCorrector.h"
-
-DEFINE_EDM_PLUGIN(TrackingRegionProducerFactory, HITRegionalPixelSeedGenerator, "HITRegionalPixelSeedGenerator");
-//
-DEFINE_FWK_MODULE(IsolatedPixelTrackCandidateProducer);
-DEFINE_FWK_MODULE(IsolatedPixelTrackCandidateL1TProducer);
-DEFINE_FWK_MODULE(IsolatedEcalPixelTrackCandidateProducer);
-DEFINE_FWK_MODULE(EcalIsolatedParticleCandidateProducer);
-DEFINE_FWK_MODULE(SiStripRegFEDSelector);
-DEFINE_FWK_MODULE(ECALRegFEDSelector);
-DEFINE_FWK_MODULE(SubdetFEDSelector);
-DEFINE_FWK_MODULE(IPTCorrector);
